extract theme config schema in separate file + rename navbar links to navbar items
